updates application for desktop use
