+++ package.json
@@ -22,6 +22,7 @@
     "react-bootstrap": "^2.5.0",
     "react-date-range": "^1.4.0",
     "react-dom": "^18.2.0",
+    "react-responsive": "^9.0.0",
     "react-router-dom": "^6.3.0",
     "react-scripts": "^5.0.1",
     "recharts": "^2.1.16",
